Removed robot-config and retained Lemlib initial structure
