--- include/robot-config.hpp
@@ -1,40 +1,0 @@
-#ifndef ROBOT_CONFIG_HPP
-#define ROBOT_CONFIG_HPP
-
-#include "main.h"
-#include "lemlib/api.hpp"
-
-extern pros::Controller controller;
-
-// motor groups
-extern pros::MotorGroup leftMotors;
-extern pros::MotorGroup rightMotors;
-
-// Inertial Sensor
-extern pros::Imu imu;
-
-// drivetrain settings
-extern lemlib::Drivetrain drivetrain;
-
-// lateral motion controller
-extern lemlib::ControllerSettings linearController;
-
-// angular motion controller
-extern lemlib::ControllerSettings angularController;
-
-// input curve for throttle input during driver control
-extern lemlib::ExpoDriveCurve throttleCurve;
-
-// input curve for steer input during driver control
-extern lemlib::ExpoDriveCurve steerCurve;
-
-// sensors for odometry
-extern lemlib::OdomSensors sensors;
-
-extern lemlib::Chassis chassis;
-
-void initialize();
-void disabled();
-void competition_initialize();
-
-#endif // ROBOT_CONFIG_HPP
